Obstale & Colletible spawning, Water movement
Added behavior to GameManager that spawns obstacles and collectibles based on a set percentage.
Added behavior to GameManager that moves the three water pieces to the front of the other waters, when the player passed them.
Removed the Unity Github extension, which made both of my computers crash unity all the time.
Made some changes to folder structure.