Marshmallow, Delete behind player
Added a new marhsmallow obstacle, and a new bobbing script than can be added to *anything* that makes the objects go up and down.
Changed the color scheme of the game.
Added a script to collectibles and obstacles that deletes them if they fall too far behind the player, so they can't be collected.