Animations Done & Bobbing
Animations are "done", so the players paddle when one paddles in real life. There are however some bugs that makes the players paddle when they're not paddling in real life.
I've also moved bobbing out of the "shouldMove" loop, so the bobbing can happen before the game has started (since boats always float).